Reset value in top-right search bar on search so that it is not stale
