+++ src/app/components/elements/MainSearch.tsx
@@ -1,4 +1,4 @@
-import React, {ChangeEvent, FormEvent, useRef, useState} from "react";
+import React, {ChangeEvent, FormEvent, useEffect, useRef, useState} from "react";
 import {SearchButton} from "./SearchButton";
 import {pushSearchToHistory} from "../../services/search";
 import {History} from "history";
@@ -7,6 +7,7 @@ import {Collapse, Form, FormGroup, Input, Label, Nav, Navbar, NavbarToggler, Nav
 import {SEARCH_CHAR_LENGTH_LIMIT} from "../../services/constants";
 import {isPhy} from "../../services/siteConstants";
 import classNames from "classnames";
+import {useLocation} from "react-router-dom";
 
 interface MainSearchProps {
     history: History;
@@ -17,7 +18,6 @@ const MainSearchComponent = ({history}: MainSearchProps) => {
     const [showSearchBox, setShowSearchBox] = useState(false);
 
     const searchInputRef = useRef<HTMLInputElement>(null);
-
     function doSearch(e: FormEvent<HTMLFormElement>) {
         e.preventDefault();
         if (searchText == "") {
@@ -26,6 +26,10 @@ const MainSearchComponent = ({history}: MainSearchProps) => {
             pushSearchToHistory(history, searchText, []);
         }
     }
+
+    // Clear this search field on location (i.e. search query) change - user should use the main search bar
+    const location = useLocation();
+    useEffect(() => { if (location.pathname === "/search") { setSearchText(""); }}, [location]);
 
     function setSearchTextAsValue(e: ChangeEvent<HTMLInputElement>) {
         setSearchText(e.target.value);
