+++ src/app/components/pages/Search.tsx
@@ -3,13 +3,12 @@ import {RouteComponentProps, withRouter} from "react-router-dom";
 import {useDispatch, useSelector} from "react-redux";
 import * as RS from "reactstrap";
 import {Col, Container, Form, Input, Row} from "reactstrap";
-import queryString from "query-string";
 import {fetchSearch} from "../../state/actions";
 import {ShowLoading} from "../handlers/ShowLoading";
 import {AppState} from "../../state/reducers";
 import {LinkToContentSummaryList} from "../elements/list-groups/ContentSummaryListGroupItem";
 import {DOCUMENT_TYPE, documentDescription, SEARCH_CHAR_LENGTH_LIMIT} from "../../services/constants";
-import {pushSearchToHistory, searchResultIsPublic} from "../../services/search";
+import {parseLocationSearch, pushSearchToHistory, searchResultIsPublic} from "../../services/search";
 import {TitleAndBreadcrumb} from "../elements/TitleAndBreadcrumb";
 import {shortcuts} from "../../services/searchResults";
 import {ShortcutResponse} from "../../../IsaacAppTypes";
@@ -33,18 +32,6 @@ function deitemise(item: Item<DOCUMENT_TYPE>) {
     return item.value;
 }
 
-function parseLocationSearch(search: string): [Nullable<string>, DOCUMENT_TYPE[]] {
-    const searchParsed = queryString.parse(search);
-
-    const parsedQuery = searchParsed.query || "";
-    const query = parsedQuery instanceof Array ? parsedQuery[0] : parsedQuery;
-
-    const parsedFilters = searchParsed.types || "";
-    const possibleFilters = (Array.isArray(parsedFilters) ? parsedFilters[0] || "" : parsedFilters || "").split(",");
-    const filters = possibleFilters.filter(pf => Object.values(DOCUMENT_TYPE).includes(pf as DOCUMENT_TYPE)) as DOCUMENT_TYPE[]
-
-    return [query, filters];
-}
 
 const selectStyle: StylesConfig<Item<DOCUMENT_TYPE>, true, GroupBase<Item<DOCUMENT_TYPE>>> = {
     multiValue: (styles: CSSObjectWithLabel) => ({...styles, backgroundColor: siteSpecific("rgba(254, 161, 0, 0.9)", "rgba(255, 181, 63, 0.9)")}),
